1、退费关联缴费信息 2、新增费用编辑费用关联流水

+++ nmt-education/src/main/java/com/nmt/education/service/course/registeration/RegistrationExpenseDetailPoMapper.java
@@ -10,7 +10,7 @@ import java.util.List;
 interface RegistrationExpenseDetailPoMapper {
     int insertSelective(RegistrationExpenseDetailPo record);
 
-    RegistrationExpenseDetailPo selectByPrimaryKey(Integer id);
+    RegistrationExpenseDetailPo selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(RegistrationExpenseDetailPo record);
